refactoring add function
