+++ camper_app/src/App.js
@@ -37,8 +37,8 @@ export function App() {
   return (
     <>
       <BrowserRouter>
-        <NavBar />
         <NotificationContainer />
+        <NavBar />
 
         <Routes>
           <Route path="/" element={<Home />} />
